feat: rename _tests_ to __tests__
